requirements.txt
# Django und Erweiterungen
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-allauth>=0.54.0
django-filter>=23.1

# Datenbank
psycopg2-binary>=2.9.6  # Für PostgreSQL (optional)

# Webserver
gunicorn>=20.1.0  # Für Produktionsumgebungen

# Für Profilbild-Downloader
requests>=2.28.2
beautifulsoup4>=4.12.0
lxml>=4.9.2
Pillow>=9.5.0  # Für Bildverarbeitung

# Für Selenium-Unterstützung (optional)
selenium>=4.9.0
webdriver-manager>=3.8.6

# Utilities
python-dotenv>=1.0.0  # Für Umgebungsvariablen
tqdm>=4.65.0  # Für Fortschrittsbalken
